refactor: 스터디 채널 조회 관련 기능 분리 - 조회 기능을 StudyChannelQueryService로 이동 - 조회 기능 이동으로 인한 변경사항 수정   - StudyChannelController 에 StudyChannelQueryService 를 추가   - 테스트 코드 역시 동일하게 변경
