+++ src/main/java/com/tenten/studybadge/study/channel/service/StudyChannelService.java
@@ -78,118 +78,6 @@ public class StudyChannelService {
         return studyChannel.getId();
     }
 
-    public StudyChannelListResponse getStudyChannels(Pageable pageable, SearchCondition searchCondition) {
-
-        Specification<StudyChannel> spec = Specification.where(null);
-
-        if (searchCondition.getStatus() != null) {
-            spec = spec.and(StudyChannelSpecification.withRecruitmentStatus(searchCondition.getStatus().name()));
-        }
-        if (searchCondition.getCategory() != null) {
-            spec = spec.and(StudyChannelSpecification.withCategory(searchCondition.getCategory().name()));
-        }
-        if (searchCondition.getType() != null) {
-            spec = spec.and(StudyChannelSpecification.withMeetingType(searchCondition.getType().name()));
-        }
-
-        Page<StudyChannel> studyChannels = studyChannelRepository.findAll(spec, pageable);
-        List<Long> ids = studyChannels.getContent().stream()
-                .map(StudyChannel::getId)
-                .toList();
-
-        List<StudyMember> leaders = studyMemberRepository.findAllWithLeader(ids, StudyMemberRole.LEADER);
-
-        Map<Long, StudyMember> leaderMap = leaders.stream()
-                .collect(Collectors.toMap(studyMember -> studyMember.getStudyChannel().getId(), Function.identity()));
-
-        return StudyChannelListResponse.from(studyChannels, leaderMap);
-    }
-
-    public StudyChannelDetailsResponse getStudyChannel(Long studyChannelId, @Nullable Long memberId) {
-        increaseViewCnt(studyChannelId);
-        StudyChannel studyChannel = studyChannelRepository.findByIdWithMember(studyChannelId).orElseThrow(NotFoundStudyChannelException::new);
-
-        LocalDate now = LocalDate.now();
-        StudyChannelDetailsResponse.StudyChannelDetailsResponseBuilder builder = createDefaultResponseBuilder(studyChannel);
-
-        if (memberId == null) {
-            return responseForAnonymousMember(builder, studyChannel, now);
-        } else {
-            Member member = memberRepository.findById(memberId).orElseThrow(NotFoundMemberException::new);
-            return responseForAuthMember(builder, studyChannel, member, now);
-        }
-
-    }
-
-    private void increaseViewCnt(Long studyChannelId) {
-        // 문제점: 기본 JPA(SimpleJpaRepository) 쿼리 메서드의 경우 exists 시 count(*) 쿼리 사용
-        if (!studyChannelRepository.existsById(studyChannelId)) {
-            throw new NotFoundStudyChannelException();
-        }
-        studyChannelRepository.increaseViewCnt(studyChannelId);
-    }
-
-    private StudyChannelDetailsResponse responseForAuthMember(
-            StudyChannelDetailsResponse.StudyChannelDetailsResponseBuilder builder,
-            StudyChannel studyChannel,
-            Member member,
-            LocalDate date) {
-
-        if (!studyChannel.isStudyMember(member.getId())) {
-            return builder.isStudyMember(false)
-                    .chattingUrl(null)
-                    .isStudyEnd(studyChannel.isStudyEnd(date))
-                    .build();
-        }
-
-        builder.isStudyMember(true)
-                .chattingUrl(studyChannel.getChattingUrl())
-                .isStudyEnd(studyChannel.isStudyEnd(date))
-                .isLeader(studyChannel.isLeader(member));
-
-        if (studyChannel.isStudyEnd(date)) {
-            StudyChannelDeposit deposit = studyChannelDepositRepository.findByStudyChannelIdAndMemberId(studyChannel.getId(), member.getId())
-                    .orElseThrow(NotFoundStudyChannelDepositException::new);
-
-            return builder.memberName(member.getName())
-                    .attendanceRatio(deposit.getAttendanceRatio())
-                    .refundsAmount(deposit.getRefundsAmount())
-                    .build();
-        } else {
-            return builder.build();
-        }
-    }
-
-    private StudyChannelDetailsResponse responseForAnonymousMember(
-            StudyChannelDetailsResponse.StudyChannelDetailsResponseBuilder builder,
-            StudyChannel studyChannel,
-            LocalDate date) {
-        return builder.isStudyMember(false)
-                .chattingUrl(null)
-                .isStudyEnd(studyChannel.isStudyEnd(date))
-                .build();
-    }
-
-    private StudyChannelDetailsResponse.StudyChannelDetailsResponseBuilder createDefaultResponseBuilder(StudyChannel studyChannel) {
-        StudyMember leader = studyChannel.getLeader();
-        StudyMember subLeader = studyChannel.getSubLeader();
-        return StudyChannelDetailsResponse.builder()
-                .studyChannelId(studyChannel.getId())
-                .studyChannelName(studyChannel.getName())
-                .studyChannelDescription(studyChannel.getDescription())
-                .deposit(studyChannel.getDeposit())
-                .category(studyChannel.getCategory())
-                .meetingType(studyChannel.getMeetingType())
-                .region(studyChannel.getRegion())
-                .startDate(studyChannel.getStudyDuration().getStudyStartDate())
-                .endDate(studyChannel.getStudyDuration().getStudyEndDate())
-                .capacity(studyChannel.getRecruitment().getRecruitmentNumber())
-                .recruitmentStatus(studyChannel.getRecruitment().getRecruitmentStatus())
-                .viewCnt(studyChannel.getViewCnt())
-                .leaderName(leader.getMember().getName())
-                .subLeaderName(Objects.requireNonNullElse(subLeader, leader).getMember().getName());
-    }
-
     public boolean checkStudyMemberInStudyChannel(Long memberId, Long studyChannelId) {
         Integer result = studyMemberRepository.existsByMemberIdAndStudyChannelIdAndStudyMemberStatus(memberId, studyChannelId);
         return result != null && result == 1;
@@ -271,19 +159,4 @@ public class StudyChannelService {
         return point;
     }
 
-    public static class StudyChannelSpecification {
-
-        public static Specification<StudyChannel> withRecruitmentStatus(String recruitmentStatus) {
-            return (root, query, cb) -> cb.equal(root.get("recruitment").get("recruitmentStatus"), recruitmentStatus);
-        }
-
-        public static Specification<StudyChannel> withCategory(String category) {
-            return (root, query, cb) -> cb.equal(root.get("category"), category);
-        }
-
-        public static Specification<StudyChannel> withMeetingType(String meetingType) {
-            return (root, query, cb) -> cb.equal(root.get("meetingType"), meetingType);
-        }
-
-    }
 }
